Relesed TaskHandler
