+++ src/ru/practicum/api/adapter/DurationAdapter.java
@@ -1,0 +1,29 @@
+package ru.practicum.api.adapter;
+
+import com.google.gson.TypeAdapter;
+import com.google.gson.stream.JsonReader;
+import com.google.gson.stream.JsonWriter;
+
+import java.io.IOException;
+import java.time.Duration;
+
+public class DurationAdapter extends TypeAdapter<Duration> {
+
+    @Override
+    public void write(JsonWriter jsonWriter, Duration duration) throws IOException {
+        if (duration == null)
+            jsonWriter.value("null");
+        else
+            jsonWriter.value(duration.toString());
+    }
+
+    @Override
+    public Duration read(JsonReader jsonReader) throws IOException {
+        String value = jsonReader.nextString();
+
+        if (value.equals("null"))
+            return null;
+        else
+            return Duration.parse(value);
+    }
+}
