+++ src/ru/practicum/api/adapter/LocalDateTimeAdapter.java
@@ -1,0 +1,23 @@
+package ru.practicum.api.adapter;
+
+import com.google.gson.TypeAdapter;
+import com.google.gson.stream.JsonReader;
+import com.google.gson.stream.JsonWriter;
+
+import java.io.IOException;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+public class LocalDateTimeAdapter extends TypeAdapter<LocalDateTime> {
+    private final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd.MM.yyyy HH:mm");
+
+    @Override
+    public void write(JsonWriter jsonWriter, LocalDateTime localDateTime) throws IOException {
+        jsonWriter.value(localDateTime.format(formatter));
+    }
+
+    @Override
+    public LocalDateTime read(JsonReader jsonReader) throws IOException {
+        return LocalDateTime.parse(jsonReader.nextString(), formatter);
+    }
+}
